+++ core/HoistProps.ts
@@ -94,17 +94,17 @@ export interface TestSupportProps {
 }
 
 export interface LayoutProps {
-    margin?: string | number | true;
-    marginTop?: string | number | true;
-    marginRight?: string | number | true;
-    marginBottom?: string | number | true;
-    marginLeft?: string | number | true;
+    margin?: string | number | boolean;
+    marginTop?: string | number | boolean;
+    marginRight?: string | number | boolean;
+    marginBottom?: string | number | boolean;
+    marginLeft?: string | number | boolean;
 
-    padding?: string | number | true;
-    paddingTop?: string | number | true;
-    paddingRight?: string | number | true;
-    paddingBottom?: string | number | true;
-    paddingLeft?: string | number | true;
+    padding?: string | number | boolean;
+    paddingTop?: string | number | boolean;
+    paddingRight?: string | number | boolean;
+    paddingBottom?: string | number | boolean;
+    paddingLeft?: string | number | boolean;
 
     height?: string | number;
     minHeight?: string | number;
@@ -119,7 +119,7 @@ export interface LayoutProps {
     flexGrow?: string | number;
     flexShrink?: string | number;
     flexWrap?: Property.FlexWrap;
-    gap?: string | number | true;
+    gap?: string | number | boolean;
 
     alignItems?: string;
     alignSelf?: string;
@@ -137,7 +137,7 @@ export interface LayoutProps {
     display?: string;
 }
 
-/** LayoutProps after resolution by `getLayoutProps()`, with `true` values resolved to strings. */
+/** LayoutProps after resolution by `getLayoutProps()`, with boolean values resolved. */
 export type ResolvedLayoutProps = {
-    [K in keyof LayoutProps]: Exclude<LayoutProps[K], true>;
+    [K in keyof LayoutProps]: Exclude<LayoutProps[K], boolean>;
 };